Move CSI metric to right domain (#2309)
* move csi to right domain
* fix docstrings
* move docs
