--- docs/source/image/critical_success_index.rst
@@ -1,22 +1,0 @@
-.. customcarditem::
-   :header: Critical Success Index (CSI)
-   :image: https://pl-flash-data.s3.amazonaws.com/assets/thumbnails/image_classification.svg
-   :tags: Image
-
-.. include:: ../links.rst
-
-############################
-Critical Success Index (CSI)
-############################
-
-Module Interface
-________________
-
-.. autoclass:: torchmetrics.image.CriticalSuccessIndex
-    :exclude-members: update, compute
-
-
-Functional Interface
-____________________
-
-.. autofunction:: torchmetrics.functional.image.critical_success_index
