+++ docs/source/regression/critical_success_index.rst
@@ -1,0 +1,22 @@
+.. customcarditem::
+   :header: Critical Success Index (CSI)
+   :image: https://pl-flash-data.s3.amazonaws.com/assets/thumbnails/image_classification.svg
+   :tags: Regression
+
+.. include:: ../links.rst
+
+############################
+Critical Success Index (CSI)
+############################
+
+Module Interface
+________________
+
+.. autoclass:: torchmetrics.regression.CriticalSuccessIndex
+    :exclude-members: update, compute
+
+
+Functional Interface
+____________________
+
+.. autofunction:: torchmetrics.functional.regression.critical_success_index
