+++ src/torchmetrics/__init__.py
@@ -95,6 +95,7 @@ from torchmetrics.nominal import (  # noqa: E402
 from torchmetrics.regression import (  # noqa: E402
     ConcordanceCorrCoef,
     CosineSimilarity,
+    CriticalSuccessIndex,
     ExplainedVariance,
     KendallRankCorrCoef,
     KLDivergence,
@@ -164,6 +165,7 @@ __all__ = [
     "ConfusionMatrix",
     "CosineSimilarity",
     "CramersV",
+    "CriticalSuccessIndex",
     "Dice",
     "TweedieDevianceScore",
     "ErrorRelativeGlobalDimensionlessSynthesis",
