+++ src/torchmetrics/functional/regression/__init__.py
@@ -11,9 +11,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 from torchmetrics.functional.regression.concordance import concordance_corrcoef
 from torchmetrics.functional.regression.cosine_similarity import cosine_similarity
+from torchmetrics.functional.regression.csi import critical_success_index
 from torchmetrics.functional.regression.explained_variance import explained_variance
 from torchmetrics.functional.regression.kendall import kendall_rank_corrcoef
 from torchmetrics.functional.regression.kl_divergence import kl_divergence
@@ -34,6 +34,7 @@ from torchmetrics.functional.regression.wmape import weighted_mean_absolute_perc
 __all__ = [
     "concordance_corrcoef",
     "cosine_similarity",
+    "critical_success_index",
     "explained_variance",
     "kendall_rank_corrcoef",
     "kl_divergence",
