+++ src/torchmetrics/regression/__init__.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 from torchmetrics.regression.concordance import ConcordanceCorrCoef
 from torchmetrics.regression.cosine_similarity import CosineSimilarity
+from torchmetrics.regression.csi import CriticalSuccessIndex
 from torchmetrics.regression.explained_variance import ExplainedVariance
 from torchmetrics.regression.kendall import KendallRankCorrCoef
 from torchmetrics.regression.kl_divergence import KLDivergence
@@ -33,6 +34,7 @@ from torchmetrics.regression.wmape import WeightedMeanAbsolutePercentageError
 __all__ = [
     "ConcordanceCorrCoef",
     "CosineSimilarity",
+    "CriticalSuccessIndex",
     "ExplainedVariance",
     "KendallRankCorrCoef",
     "KLDivergence",
